+++ packages/contracts/contracts/libs/v0.8.x/minter-libs/DALinLib.sol
@@ -13,14 +13,13 @@ pragma solidity ^0.8.0;
 
 library DALinLib {
     struct DAProjectConfig {
-        // @dev max uint64 ~= 1.8e19 sec ~= 570 billion years
-        uint64 timestampStart;
-        uint64 timestampEnd;
-        // @dev Prices are packed internally as uint128, resulting in a maximum
-        // allowed price of ~3.4e20 ETH. This is many orders of magnitude
-        // greater than current ETH supply.
-        uint128 startPrice;
-        uint128 basePrice;
+        // @dev max uint40 ~= 1.1e12 sec ~= 34 thousand years
+        uint40 timestampStart;
+        uint40 timestampEnd;
+        // @dev max uint88 ~= 3e26 Wei = ~300 million ETH, which is well above
+        // the expected prices of any NFT mint in the foreseeable future.
+        uint88 startPrice;
+        uint88 basePrice;
     }
 
     /**
@@ -45,10 +44,10 @@ library DALinLib {
      */
     function setAuctionDetailsLin(
         DAProjectConfig storage _DAProjectConfig,
-        uint64 _auctionTimestampStart,
-        uint64 _auctionTimestampEnd,
-        uint128 _startPrice,
-        uint128 _basePrice,
+        uint40 _auctionTimestampStart,
+        uint40 _auctionTimestampEnd,
+        uint88 _startPrice,
+        uint88 _basePrice,
         bool _allowReconfigureAfterStart
     ) internal {
         require(
