+++ packages/contracts/contracts/minter-suite/Minters/MinterDALinHolderV5.sol
@@ -397,10 +397,10 @@ contract MinterDALinHolderV5 is
     function setAuctionDetails(
         uint256 _projectId,
         address _coreContract,
-        uint64 _auctionTimestampStart,
-        uint64 _auctionTimestampEnd,
-        uint128 _startPrice,
-        uint128 _basePrice
+        uint40 _auctionTimestampStart,
+        uint40 _auctionTimestampEnd,
+        uint88 _startPrice,
+        uint88 _basePrice
     ) external {
         AuthLib.onlyArtist({
             _projectId: _projectId,
@@ -597,10 +597,10 @@ contract MinterDALinHolderV5 is
         external
         view
         returns (
-            uint64 timestampStart,
-            uint64 timestampEnd,
-            uint128 startPrice,
-            uint128 basePrice
+            uint40 timestampStart,
+            uint40 timestampEnd,
+            uint88 startPrice,
+            uint88 basePrice
         )
     {
         DALinLib.DAProjectConfig
